+++ src/data/index.js
@@ -1,5 +1,6 @@
 import courses from "./Courses";
 import digitalTwin from "./digitalTwin";
+import FormatUtils from "./format.utils";
 import jobs from "./Jobs";
 import SkillsUtils from "./skills.utils";
 
@@ -9,6 +10,7 @@ const data = {
     digitalTwin,
     utils: { 
         skills : SkillsUtils,
+        format: FormatUtils,
     },
 }