+++ src/visualizers/builder.js
@@ -1,4 +1,5 @@
 const courses = require("../data/Courses");
+const { detectFormat } = require("../data/format.utils");
 const jobs = require("../data/Jobs");
 const { CourseVisualizer } = require("./list/CourseVisualizer");
 const { JobVisualizer } = require("./list/JobVisualizer");
@@ -64,13 +65,16 @@ function buildVisualization(visualInfo){
 }
 
 function buildRulesForSingleMap(params){
-    const jsonUrl = params['jsonUrl'];
+    const { jsonUrl , data } = params;
+    delete params.jsonUrl;
+    delete params.data;
 
     const rules = {
         visuals: [
             {
                 type: 'HexagonMap',
                 url: jsonUrl,
+                data: data,
                 properties: params,
             }
         ],
@@ -81,9 +85,54 @@ function buildRulesForSingleMap(params){
     return rules;
 }
 
+function getRulesFromSignals(json){
+    for (let i = 0; i < json.data.length; i++) {
+        const element = json.data[i];
+        element.type = "HexagonMap";
+    }
+
+    const rules = {
+        "visuals": json.data,
+        "properties": {
+            "timeLabels": json.info.timeLabels,
+        }
+    }
+
+    return rules;
+}
+
+async function buildRules(params){
+    const { jsonUrl } = params;
+    try {
+        const json = await d3.json(jsonUrl);
+        const format = detectFormat(json);
+
+        params.data = json;
+        delete params.jsonUrl;
+
+        if (format == 'digitalTwin'){
+            return buildRulesForSingleMap(params);
+        }
+
+        if (format == 'signals'){
+            return getRulesFromSignals(json);
+        }
+
+        if (format == 'rules') return json;
+        // Else
+        return {visuals:[], params:{}};
+
+    } catch (error) {
+        console.log(error);
+        return {visuals:[], params:{}};
+    }
+    
+}
+
 const builder = {
     buildVisualization,
     buildRulesForSingleMap,
+    buildRules,
 }
 
 module.exports = builder;